bug(dropdown): Adjustments for context data
