+++ packages/app/components/PostList/PostList.tsx
@@ -2,7 +2,7 @@ import { styled } from "@/stitches.config";
 import { useState } from "react";
 import Post from "../Post";
 import {v4 as uuidV4} from "uuid";
-import { PostListProps } from "@/types";
+import PostListProps from "@/types";
 
 
 const StyledPostList = styled("div", {
@@ -11,12 +11,14 @@ const StyledPostList = styled("div", {
   rowGap: "$4",
 });
 
-const PostList = ({ posts, sort }: PostListProps) => {
+
+
+const PostList = (props) => {
+
   return (
     <StyledPostList>
-      {sort === "newest"
-        ? posts.sort((a, b) => a.submissionDate - b.submissionDate).map((post) => <Post key={uuidV4()} {...post} />)
-        : posts.sort((a, b) => a.numberOfUpvotes - b.numberOfUpvotes).map((post) => <Post key={uuidV4()} {...post} />}
+      {props.sort === "newest" && props.posts.sort((a, b) => a.submissionDate - b.submissionDate).map((post) => <Post key={uuidV4()} {...post} />)}
+      {props.sort === "trending" && props.posts.sort((a, b) => a.numberOfUpvotes - b.numberOfUpvotes).map((post) => <Post key={uuidV4()} {...post} />)}
     </StyledPostList>
   );
 };
